requirements.txt
# Core ML and data processing libraries
tensorflow>=2.8.0
opencv-python>=4.5.5
numpy>=1.22.3
pandas>=1.4.2
scikit-learn>=1.0.2
scipy>=1.8.0

# Web server and API
flask>=2.1.1
flask-cors>=3.0.10
gunicorn>=20.1.0
flask_socketio

# MongoDB
pymongo>=4.1.1

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
pillow>=9.1.0
matplotlib>=3.5.1